update:修改sdk中用于队列排序的level类型，从int变为level，这样，就可使之基于时间也是可行的

+++ messagesdk/src/main/java/com/why/message/Message.java
@@ -32,7 +32,7 @@ public class Message {
 	boolean used = false;
 
 	//消息的优先级
-	public int level;
+	public long level;
 
 	//消息类型 据FLAG区分为同步消息和异步消息
 	int flag;
